2 hane olacak şekilde düzenlendi

+++ Okyanus.EntityLayer/Entities/Basket.cs
@@ -9,9 +9,9 @@ namespace Okyanus.EntityLayer.Entities
 {
     public class Basket : BaseEntity
     {
-        public double Price { get; set; }
-        public double Count { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal Count { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public Product Product { get; set; }
     }
