+++ Okyanus.EntityLayer/Entities/OrderDetail.cs
@@ -9,9 +9,9 @@ namespace Okyanus.EntityLayer.Entities
 {
     public class OrderDetail : BaseEntity
     {
-        public double Count { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Count { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
 
 
         public int ProductID { get; set; }
