+++ Okyanus.EntityLayer/Entities/ProductType.cs
@@ -10,7 +10,7 @@ namespace Okyanus.EntityLayer.Entities
     public class ProductType : BaseEntity
     {
         public string Birim { get; set; }
-        public double IncreaseAmount { get; set; }
+        public decimal IncreaseAmount { get; set; }
 
         public virtual List<Product> Products { get; set; }
     }
