+++ OkyanusWebAPI/Models/OrderDetailVM/ResultOrderDetailVM.cs
@@ -11,9 +11,9 @@ namespace OkyanusWebAPI.Models.OrderDetailVM
         public bool Status { get; set; }
 
 
-        public double Count { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Count { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public int OrderID { get; set; }
         public ResultProductVM Product { get; set; }
