+++ OkyanusWebAPI/Models/ProductTypeVM/CreateProductTypeVM.cs
@@ -3,6 +3,6 @@
     public class CreateProductTypeVM
     {
         public string Birim { get; set; }
-        public double IncreaseAmount { get; set; }
+        public decimal IncreaseAmount { get; set; }
     }
 }
