+++ OkyanusWebAPI/Models/ProductTypeVM/ResultProductTypeVM.cs
@@ -8,6 +8,6 @@
         public bool Status { get; set; }
 
         public string Birim { get; set; }
-        public double IncreaseAmount { get; set; }
+        public decimal IncreaseAmount { get; set; }
     }
 }
