+++ OkyanusWebAPI/Models/ProductVM/CreateProductVM.cs
@@ -3,11 +3,11 @@
     public class CreateProductVM
     {
         public string ProductName { get; set; }
-        public double Price { get; set; }
-        public double? DiscountedPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? DiscountedPrice { get; set; }
         public string? ImageUrl { get; set; }
         public string? Description { get; set; }
-        public int Stock { get; set; }
+        public decimal Stock { get; set; }
         public string AnaBarcode { get; set; }
