+++ OkyanusWebUI/Models/OrderDetailVM/CreateOrderDetailVM.cs
@@ -2,9 +2,9 @@
 {
     public class CreateOrderDetailVM
     {
-        public int Count { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Count { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public int OrderID { get; set; }
     }
