+++ OkyanusWebUI/Models/OrderDetailVM/ResultOrderDetailVM.cs
@@ -8,9 +8,9 @@ namespace OkyanusWebUI.Models.OrderDetailVM
         public DateTime CreatedDate { get; set; }
         public DateTime UpdatedDate { get; set; }
 
-        public double Count { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+        public decimal Count { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         public int ProductID { get; set; }
         public int OrderID { get; set; }
         public ResultProductVM Product { get; set; }
